Exercise the write_pickle_data function in testing.
